Completes ex 5.1, adds login to bloglist frontend
This commit adds login requirements to the frontend of the bloglist
project. It adds services/login.js, which handles login requests with
the database server. It also modifies services/blogs.js to add a
token variable and a function to manipulate the token variable.

Login is handled by the new doLogin function in app.js, which takes
an object with username and password as a parameter and attempts to
login with the credentials. If successful, it saves the returned user
object in state and also saves the token with the new function added
to services/blogs.js. If unsuccessful, an alert window informs the
user of this fact.

Te webpage is now rendered conditionally in app.jsx based on whether
a user has logged in. If no user is logged in, a Login component is
rendered, defined in the new components/Login.jsx file. The Login
component renders a header prompting login and a login form. The
username and password entered into the form are saved in state
variables defined in the Login component. The doLogin function is
passed to this component, and is called with the Login state
variables for username and password in the event handler function
for the form's submission.

If a user has logged in, the app.jsx conditionally renders a Bloglist
component defined in components/Bloglist.jsx. This is essentially
the same as the previous commit render, although it adds a <p> with
the name of the user logged in. This component is passed the list of
blogs and the user object which are saved in app.jsx state variables.

Certain additional changes were made to .json and ignore files to
enable this frontend to work with the backend developed in FSO part
4. As FSO has switched from create-react-app to vite while I was
doing part 4, I was unfamiliar with some of the changed requirements
for react (most notably, storing proxy in vite.config.js rather than
in package.json). I've gone through the changed material and should
have a better handle on the differences going forward.