Completes ex. 5.2, saves login to local storage
This commit saves login to local storage, so that when a user logs in
this remains after the page is reloaded.

This commit also adds a logout button which is rendered in the
Bloglist component next to the user's name. Logout functionality is
handled in a doLogout function which is defined in app.jsx and passed
to the Bloglist component.