Completes ex. 5.3, enables adding new blogs
This commit adds a Newblog component in components/Newblog.jsx which
renders a form to add a new blog entry. This form only appears when a
user is logged in, and uses the token from that login to save the
blog to the database. The blog is added through a function defined in
app.jsx and passed to the Newblog component. The Newblog component
contains state variables for the blog's title, author, and url, which
are used to make a new blog object which is passed as a parameter to
the addBlog function.

A new function to POST data to the server is also added to
services/blogs.js

If the POST to the database fails for any reason, the addBlog
function displays the error message in an alert. When a blog is
successfully added the Newblog component entry fields are reset to
blank, but if it fails they are left as they were.

Because of the exmaple images for how the page should appear in the
FSO exercise, this commit also makes some slight modifications to how
the page is rendered. When a user is logged in, App.jsx now
conditionally renders the header, username, and logout button itself,
rather than having the Bloglist component render those. Bloglist is
now only responsible for rendering the list of blogs. Accordingly, it
is now passed only the blog state variable, not the user variable or
the logout function.