Completes ex 5.4, adds notification messages
This commit adds a Notification component, which is rendered beneath
the header on the page. The Notification component is passed an
object which consists of a string content and a color, and renders a
box using that color as outline and font containing the string. If
the object is null, nothing is rendered. Notifications timeout after
5 seconds, and only the most recent message is rendered - useRef is
used to keep track of whether to apply a given timeout to set the
Notification component to null. Notifications are displayed on login,
logout, and adding a blog - green is used for successful operations,
red for unsuccessful.

As the images in the exercise show the Notification component between
the header and the data entry form, this commit also modifies the
Login component so that the header associated with it is now rendered
in the App component. The header content is now conditionally
rendered based on the login state.