Completes ex 5.5, makes new blog form togglable
This commit creates a new Togglable component, which adds a button
that makes its child components visible or not. Togglable takes as
props the text for the button to make its children visible (making
not visible is hardcoded as 'cancel'), and a useRef variable. The
Togglable component passes back via useRef the function to switch
whether children are or are not visible, which is used to hide the
form after a successful blog add.

NB - Exercise 5.6 was already completed in previous commits.