Completes ex. 5.7, adds view/hide more blog info
NB - exercise 5.6 was unknowingly completed prior to exercise 5.4

This commit adds a button to each blog entry which allows the viewing
or hiding of additional blog info (url, likes, user who added the
blog to the bloglist). It also adds a border around the blog. Due to
the intended presentation/layout of the additions, this is not done
using the Togglable component, but is done with state variables in
the Blog component (this is specified in the exercise as the intended
way to complete it).