Completes ex 5.8, adds user to newly created blog
This commit fixes an issue where a newly created blog would not
display the username of the user who created it. This was because the
information returned by the POST operation included only the user's
id, but not other associated information stored in the users
database. To resolve, this commit adds to the addBlog function in
App.jsx, adding the username and name stored in the User state
variable to the Blog object returned by the POST operation - because
the POST uses the token stored in the user variable, the associated
username and name must be correct for the new blog. This also avoids
an additional call to the database to get the info.

This commit also removes a stray console.log in Blog.jsx.