Completes ex. 5.9, enables add like buttons
This commit adds functionality to the "like" buttons in the
viewable/hideable blog info. Clicking the button increments the like
total for that blog by 1 and updates the display accordingly. The
PUT call to the database is defined in services/blogs.js as
updateBlog.

The function which calls the PUT request is defined in App.jsx as
addLike - it takes a blog's id as a parameter, finds the blog in the
blogs state variable, creates a new blog object identical to the
found blog with likes incremented by 1, PUTs the new blog, and
replaces the blog in the state variable with the new incremented
blog.

The addLike function is passed as a prop to the Bloglist component.
The Bloglist component passes to each Blog component a function which
calls addLike with the blog's id - that function is then set as the
onClick for the like button.

To enable this commit, the PUT command in the backend is updated to
remove the requirement for a valid token from the creating user.