Completes ex. 5.11, adds blog remove buttons
This commit adds remove buttons to the viewable/hideable blog info
sections. The remove button, when clicked, prompts the user to
confirm, then sends a DELETE command to the database server and
removes the blog from local storage.

The delete command is handled in services/blogs.js with the new
deleteBlog function. This takes the id of the blog to be deleted as
a parameter and sends the request to remove the blog with the current
token.

The delete command is called by the removeBlog function in App.jsx,
which takes an id as a parameter, prompts the user to confirm
removal, and calls the deleteBlog function. The App component passes
this function as a prop to the Bloglist component.

The delete buttons only appear if the logged in user is the user who
added the blog to the database (and thus has authority to remove it).
This is done by App.jsx passing the username of the current user as a
prop to the Bloglist component. When rendering the Blog components,
the Bloglist component compares the username to that blog's
user.username member. If they are the same, it passes as a prop a
function that calls deleteBlog with the blog's id; if different, it
passes null. The Blog component conditionally renders the button if
this prop is not null, and uses the prop function as the onClick.