Completes ex. 5.13, tests initial Blog render
This commit adds necessary libraries for jest testing, and adds a
test for the Blog component to ensure its initial render displays the
name and author of the blog but hides the url and number of likes.

This commit also modifies the Note component so that the hidden items
are hidden through changing style rather than through conditional
rendering, and adds class identifiers to the <div>'s that are hidden
at the initial render.