Completes ex. 5.16, tests new blog form
This commit adds tests for the Newblog component. It tests that the
function passed as a prop is called on clicking the form submission
button, and that the submitted data is correct.

This commit also makes changes to the Newblog component, adding id
fields to the text entry elements of the form.