Completes ex. 5.18, tests login e2e, adds pun
This commit adds two tests to the cypress test to perform end-to-end
testing on the login functionality. This validates that correct
credentials succeed at logging in, display the correct and correctly
formatted notification, and display the bloglist. It further tests
that incorrect credentials display the correct and correctly
formatted notification and continues to display the login screen.

This commit also updates the login message to prompt the user to
"(b)log in to application" and upon success states that the user has
"(b)logged in" (see what I did there?)