Completes ex. 5.19-P, tests adding new blogs
This commit adds a test to make sure a logged-in user can add a new
blog. It uses Playwright to create a new blog, and verifies that the
blog appears in the list of blogs. It also updates the Newblog component
so that it has data-testid attributes for its text entry fields.

This commit pulls out the functionality to login and create blogs into
testHelper.js as these actions are or will be repeated in multiple test
and beforeEach blocks.