Completes ex 5.20-P, tests that blog can be liked
This commit adds a Playwright test to verify that a logged-in user can
like a blog and that clicking the like button increments the total
likes.

This also adds a logout helper function and a new describe section not
yet used - these will be necessary for future exercises, but not yet - I
wrote them before I realized I did not yet need them.