Completes ex 5.23-P, tests blog order by likes
This commit adds Playwright testing of blog ordering in the bloglist,
validating that it orders and re-orders the blogs in descending order of
likes.

This also increases timeout from 3 seconds to 6 seconds, as occasionally
the tests were timing out waiting for server response. After increasing
timeout this is no longer observed.